feat: implement DSA sheet validation and parsing logic; add new validation utilities and integrate with API routes for improved data handling
