+++ apps/api/package.json
@@ -33,7 +33,8 @@
     "raw-body": "^3.0.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "uuid": "^9.0.0"
+    "uuid": "^9.0.0",
+    "zod": "^3.25.76"
   },
   "devDependencies": {
     "@tbe/eslint-config": "workspace:*",
