+++ apps/dsayatra/src/pages/404.tsx
@@ -1,15 +1,3 @@
-import Link from "next/link";
+import { DsaYatraNotFoundPage } from "@tbe/components";
 
-export default function NotFound() {
-  return (
-    <div className="min-h-screen flex items-center justify-center bg-gray-100">
-      <div className="text-center">
-        <h1 className="text-4xl font-bold mb-4">404</h1>
-        <p className="text-xl text-gray-600 mb-4">Oops! Page not found</p>
-        <Link href="/" className="text-blue-500 hover:text-blue-700 underline">
-          Return to Home
-        </Link>
-      </div>
-    </div>
-  );
-}
+export default DsaYatraNotFoundPage;
