+++ apps/oncampus/public/sitemap.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://oncampus.theboringeducation.com</loc><lastmod>2026-04-09T06:23:22.149Z</lastmod><changefreq>weekly</changefreq><priority>1</priority></url>
-<url><loc>https://oncampus.theboringeducation.com/auth/callback</loc><lastmod>2026-04-09T06:23:22.157Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://oncampus.theboringeducation.com/campus-prep</loc><lastmod>2026-04-09T06:23:22.157Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://oncampus.theboringeducation.com/dashboard</loc><lastmod>2026-04-09T06:23:22.157Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
-<url><loc>https://oncampus.theboringeducation.com/login</loc><lastmod>2026-04-09T06:23:22.157Z</lastmod><changefreq>monthly</changefreq><priority>0.5</priority></url>
+<url><loc>https://oncampus.theboringeducation.com</loc><lastmod>2026-04-09T10:40:27.864Z</lastmod><changefreq>weekly</changefreq><priority>1</priority></url>
+<url><loc>https://oncampus.theboringeducation.com/auth/callback</loc><lastmod>2026-04-09T10:40:27.903Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://oncampus.theboringeducation.com/campus-prep</loc><lastmod>2026-04-09T10:40:27.903Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://oncampus.theboringeducation.com/dashboard</loc><lastmod>2026-04-09T10:40:27.903Z</lastmod><changefreq>weekly</changefreq><priority>0.7</priority></url>
+<url><loc>https://oncampus.theboringeducation.com/login</loc><lastmod>2026-04-09T10:40:27.903Z</lastmod><changefreq>monthly</changefreq><priority>0.5</priority></url>
 </urlset>
