+++ apps/oncampus/src/pages/index.tsx
@@ -1,14 +1,3 @@
-import { Footer, Navbar } from "@tbe/components";
-import { Fragment } from "react";
+import { OnCampusLandingPage } from "@tbe/components";
 
-import CampusPrepLanding from "./campus-prep";
-
-export default function Home() {
-  return (
-    <Fragment>
-      <Navbar variant="oncampus" theme="dark" />
-      <CampusPrepLanding />
-      <Footer variant="oncampus" />
-    </Fragment>
-  );
-}
+export default OnCampusLandingPage;
