+++ apps/testing/src/unit/api-routes/auth-logout.test.ts
@@ -1,0 +1,46 @@
+import type { NextApiRequest, NextApiResponse } from "next";
+import { createMocks } from "node-mocks-http";
+import { describe, expect, it, vi } from "vitest";
+
+vi.mock("../../../../api/src/lib/utils", () => ({
+  sendAPIResponse: (payload: unknown) => payload,
+}));
+
+import handler from "../../../../api/src/pages/api/v1/auth/logout";
+
+describe("Auth logout API route", () => {
+  it("returns 204 for OPTIONS", async () => {
+    const { req, res } = createMocks<NextApiRequest, NextApiResponse>({
+      method: "OPTIONS",
+    });
+
+    await handler(req, res);
+
+    expect(res._getStatusCode()).toBe(204);
+  });
+
+  it("returns 405 for non-POST methods", async () => {
+    const { req, res } = createMocks<NextApiRequest, NextApiResponse>({
+      method: "GET",
+    });
+
+    await handler(req, res);
+
+    expect(res._getStatusCode()).toBe(405);
+    const data = JSON.parse(res._getData());
+    expect(data.message).toBe("Method not allowed");
+  });
+
+  it("returns 200 with success message for POST", async () => {
+    const { req, res } = createMocks<NextApiRequest, NextApiResponse>({
+      method: "POST",
+    });
+
+    await handler(req, res);
+
+    expect(res._getStatusCode()).toBe(200);
+    const data = JSON.parse(res._getData());
+    expect(data.status).toBe(true);
+    expect(data.message).toBe("Logged out successfully");
+  });
+});
