+++ docs/code-stability.md
@@ -40,17 +40,17 @@ Parallel work tracker: pick an unchecked `[ ]` item, branch, implement, PR, then
 
 ### 3a. Packages
 
-- [ ] **Hooks without tests**: add unit tests for `useStudyGuide`, `useDsaTopicSummaries` / `useDsaQuestionsForTopic`, `usePyGamification`, `useAdmin`, `useCertificate`, `useCashfreePayment`, `usePaymentAccess`, `useResumeEvaluation`, `useQuestionStarred`, `useNotifications`, `useDailyPrepEncouragement`, `useUnskilledGraphData`, `useSkillPlaylist`, etc. ([packages/hooks/src](../packages/hooks/src)). **Progress:** `useLeaderboard` — [useLeaderboard.test.ts](../apps/testing/src/unit/hooks/useLeaderboard.test.ts); `useGamification` — [useGamification.test.ts](../apps/testing/src/unit/hooks/useGamification.test.ts); onboarding config — [onboarding-config.test.ts](../apps/testing/src/unit/onboarding/onboarding-config.test.ts); `useOnboarding` — [useOnboarding.test.ts](../apps/testing/src/unit/hooks/useOnboarding.test.ts).
+- [ ] **Hooks without tests**: add unit tests for `useStudyGuide`, `useDsaTopicSummaries` / `useDsaQuestionsForTopic`, `useAdmin`, `useCertificate`, `useCashfreePayment`, `usePaymentAccess`, `useResumeEvaluation`, `useNotifications`, `useUnskilledGraphData`, `useSkillPlaylist`, etc. ([packages/hooks/src](../packages/hooks/src)). **Progress:** `useLeaderboard` — [useLeaderboard.test.ts](../apps/testing/src/unit/hooks/useLeaderboard.test.ts); `useGamification` — [useGamification.test.ts](../apps/testing/src/unit/hooks/useGamification.test.ts); `usePyGamification` — [usePyGamification.test.ts](../apps/testing/src/unit/hooks/usePyGamification.test.ts); `useDailyPrepEncouragement` — [useDailyPrepEncouragement.test.ts](../apps/testing/src/unit/hooks/useDailyPrepEncouragement.test.ts); `useQuestionStarred` — [useQuestionStarred.test.ts](../apps/testing/src/unit/hooks/useQuestionStarred.test.ts); onboarding config — [onboarding-config.test.ts](../apps/testing/src/unit/onboarding/onboarding-config.test.ts); `useOnboarding` — [useOnboarding.test.ts](../apps/testing/src/unit/hooks/useOnboarding.test.ts).
 - [ ] **Utils**: extend coverage for [packages/utils/src](../packages/utils/src) beyond existing quiz/onboarding/dsa tests (auth, analytics edge cases).
 
 ### 3b. Components
 
-- [ ] **High-traffic UI**: add RTL tests for critical `@tbe/components` containers (e.g. prep/workspace cards, payment flows) prioritizing regressions seen in production.
+- [ ] **High-traffic UI**: add RTL tests for critical `@tbe/components` containers (e.g. prep/workspace cards, payment flows) prioritizing regressions seen in production. **Progress (Prep Yatra pages in package):** full-page compositions live under [packages/components/src/prepyatra/pages](../packages/components/src/prepyatra/pages) (`PrepYatraLandingPage`, `PrepYatraDashboardPage`, `PrepYatraPricingPage`, `PrepYatraLoginPage`, `PrepYatraPublicJourneyPage`); app routes re-export them from [apps/prep-yatra/src/pages](../apps/prep-yatra/src/pages). **Progress (On Campus pages in package):** [packages/components/src/oncampus/pages](../packages/components/src/oncampus/pages) (`OnCampusLandingPage`, `OnCampusLoginPage`); re-exported from [apps/oncampus/src/pages](../apps/oncampus/src/pages) (`/`, `/campus-prep`, `/login`). **Progress (DSA Yatra pages in package):** [packages/components/src/dsayatra/pages](../packages/components/src/dsayatra/pages) (`DsaYatraLandingPage`, `DsaYatraLoginPage`, `DsaYatraNotFoundPage`); re-exported from [apps/dsayatra/src/pages](../apps/dsayatra/src/pages) (`/`, `/login`, `404`).
 - [ ] **Quizzes app alignment**: after Gamification dedupe, add tests that assert provider wiring in [apps/quizes](../apps/quizes).
 
 ### 3c. API (handler-level unit tests)
 
-- [ ] **Auth routes**: [apps/api/src/pages/api/v1/auth](../apps/api/src/pages/api/v1/auth) — login, token, session, refresh, logout, callbacks. **Progress:** `session` (GET bearer → user payload) — [auth-session.test.ts](../apps/testing/src/unit/api-routes/auth-session.test.ts).
+- [ ] **Auth routes**: [apps/api/src/pages/api/v1/auth](../apps/api/src/pages/api/v1/auth) — login, token, session, refresh, logout, callbacks. **Progress:** `session` — [auth-session.test.ts](../apps/testing/src/unit/api-routes/auth-session.test.ts); `logout` (POST) — [auth-logout.test.ts](../apps/testing/src/unit/api-routes/auth-logout.test.ts); `token` (POST auth code exchange) — [auth-token.test.ts](../apps/testing/src/unit/api-routes/auth-token.test.ts); `refresh` (POST) — [auth-refresh.test.ts](../apps/testing/src/unit/api-routes/auth-refresh.test.ts); `login` (GET OAuth start) — [auth-login.test.ts](../apps/testing/src/unit/api-routes/auth-login.test.ts).
 - [ ] **Admin routes**: dashboard, users, content, coupons, quiz analytics, mentorship — batch by domain.
 - [ ] **Projects / webinar / youfocus / unskilled**: greenfield tests per handler group.
 - [ ] **Interview prep** sub-routes: study-guide, aptitude, sheet CRUD, company-types, upload — fill gaps beyond [interview-prep-index](../apps/testing/src/unit/api-routes/interview-prep-index.test.ts) / [dsa-sheet](../apps/testing/src/unit/api-routes/dsa-sheet.test.ts).
