+++ packages/components/src/dsayatra/pages/DsaYatraNotFoundPage.tsx
@@ -1,0 +1,18 @@
+import Link from "next/link";
+
+/**
+ * DSA Yatra 404 page.
+ */
+export default function DsaYatraNotFoundPage() {
+  return (
+    <div className="min-h-screen flex items-center justify-center bg-gray-100">
+      <div className="text-center">
+        <h1 className="text-4xl font-bold mb-4">404</h1>
+        <p className="text-xl text-gray-600 mb-4">Oops! Page not found</p>
+        <Link href="/" className="text-blue-500 hover:text-blue-700 underline">
+          Return to Home
+        </Link>
+      </div>
+    </div>
+  );
+}
