+++ packages/components/src/dsayatra/pages/index.ts
@@ -1,0 +1,4 @@
+export type { DsaYatraLandingPageProps } from "./DsaYatraLandingPage";
+export { DsaYatraLandingPage } from "./DsaYatraLandingPage";
+export { default as DsaYatraLoginPage } from "./DsaYatraLoginPage";
+export { default as DsaYatraNotFoundPage } from "./DsaYatraNotFoundPage";
