+++ packages/components/src/oncampus/index.ts
@@ -1,0 +1,1 @@
+export * from "./pages";
