+++ packages/components/src/oncampus/pages/index.ts
@@ -1,0 +1,2 @@
+export { default as OnCampusLandingPage } from "./OnCampusLandingPage";
+export { default as OnCampusLoginPage } from "./OnCampusLoginPage";
