+++ packages/components/src/prepyatra/pages/index.ts
@@ -1,0 +1,5 @@
+export { default as PrepYatraDashboardPage } from "./PrepYatraDashboardPage";
+export { PrepYatraLandingPage } from "./PrepYatraLandingPage";
+export { default as PrepYatraLoginPage } from "./PrepYatraLoginPage";
+export { PrepYatraPricingPage } from "./PrepYatraPricingPage";
+export { default as PrepYatraPublicJourneyPage } from "./PrepYatraPublicJourneyPage";
